+++ tests/test_ip_filter.cpp
@@ -35,7 +35,7 @@ TEST(TestSimple, TestGetIpPool){
 
 TEST(TestSimple, TestShowIpPool){
 
-    std::vector <std::vector<int>> input_pool = 
+    std::vector <std::vector<uint8_t>> input_pool = 
     {
         {10, 10, 10, 10},
         {1, 2, 3, 4},
@@ -69,7 +69,7 @@ TEST(TestSimple, TestShowIpPool){
 
 TEST(TestSimple, TestReverseSort){
 
-    std::vector <std::vector<int>> input_pool = 
+    std::vector <std::vector<uint8_t>> input_pool = 
     {
         {4,3,2,1},
         {5,4,3,2},
@@ -85,7 +85,7 @@ TEST(TestSimple, TestReverseSort){
         {6,7,8,9}
     };
 
-    std::vector <std::vector<int>> expected = 
+    std::vector <std::vector<uint8_t>> expected = 
     {   
         {4,6,250,240},
         {6,7,8,9},
@@ -107,7 +107,7 @@ TEST(TestSimple, TestReverseSort){
 
 TEST(TestSimple, TestFilterByCondition){
 
-    std::vector <std::vector<int>> input_pool = 
+    std::vector <std::vector<uint8_t>> input_pool = 
     {
         {2,3,4,5},
         {1,2,3,4},
@@ -124,7 +124,7 @@ TEST(TestSimple, TestFilterByCondition){
         {1,8,7,46}
     };
 
-    const std::vector <int> some_val;
+    const std::vector <uint8_t> some_val;
     auto condition_func1 = [](decltype(some_val)& input)-> bool {return input[0]==1;};
     auto result_1 = filter_by_condition(input_pool, condition_func1);
     decltype(input_pool) expected_1 =
